+++ include/settings.h
@@ -207,7 +207,7 @@ const char* const DEFAULT_PASSWORD = "password";
 #define SERIAL_PORT stepperSerial
 
 // Match to your driver
-#define R_SENSE 0.11f
+#define R_SENSE 0.08f
 
 // Hardware pin for indicator LED *note* internal LED on esp32 Dev board is pin
 // 2
@@ -274,7 +274,7 @@ constexpr const char* ANY = "any";
 #define RUNTIMECONFIG_JSON_SIZE 1000 + DEBUG_LOG_BUFFER_SIZE
 
 // Uncomment to use guardrails for ERG mode in the stepper loop.
-// #define ERG_GUARDRAILS
+#define ERG_GUARDRAILS
 
 // Uncomment to enable the use of the power table for ERG mode.
 #define ERG_MODE_USE_POWER_TABLE
